% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparision.R, R/simulation_code.R
\docType{data}
\name{ants}
\alias{ants}
\title{Comparision of simulated data with emperical data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 1557 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 1557 columns.
}
\usage{
ants

trees <- sim.fbd.taxa(n=10, numbsim=1, lambda = 1, mu = 1, psi = 1, frac = .01)
}
\arguments{
\item{n}{Number of extant sampled tips}

\item{numbsim}{Number of trees to simulate}

\item{mu}{Extinction rate}

\item{lambda}{Vector of speciation rates, the rate in entry i is the speciation rate prior (ancestral) to time times\link{i}.}

\item{frac}{}
}
\value{
trees
}
\description{
Comparision of simulated data with emperical data

data simulation
}
\keyword{datasets}
