% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size}
\alias{effect_size}
\title{Calculation of effect size to know how many standard deviation away from the median is the observed value.}
\usage{
effect_size(datasdf = datas)
}
\arguments{
\item{datasdf}{default data frame which contains the data}
}
\value{
eff_size
}
\description{
Calculation of effect size to know how many standard deviation away from the median is the observed value.
}
