% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_data_visualization.R
\docType{data}
\name{i}
\alias{i}
\title{visualization of data as tree plot}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
i
}
\description{
visualization of data as tree plot
}
\keyword{datasets}
