% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{simulation}
\alias{simulation}
\title{data simulation}
\usage{
simulation(n=10, numbsim=1, lambda = 0.01806175, mu = 1.078013e-03 , psi = 5.172419e-02, frac = .01)
}
\arguments{
\item{n}{Number of extant sampled tips}

\item{numbsim}{Number of trees to simulate}

\item{lambda}{Vector of speciation rates, the rate in entry i is the speciation rate prior (ancestral) to time times\link{i}.}

\item{mu}{Extinction rate}

\item{psi}{Fossil sampling rate}

\item{frac}{The extant tip of tree is n/frac}
}
\value{
result
}
\description{
data simulation
}
