% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{tree}
\alias{tree}
\title{data simulation}
\usage{
i = 1
while (i <= 20) {
trees[i] <- (tree(n=10, numbsim=2, lambda[i], mu[i], psi[i], frac = .01))
 i <- i + 1}
}
\arguments{
\item{n}{Number of extant sampled tips}

\item{numbsim}{Number of trees to simulate}

\item{lambda}{Vector of speciation rates, the rate in entry i is the speciation rate prior (ancestral) to time times\link{i}.}

\item{mu}{Extinction rate}

\item{psi}{Fossil sampling rate}

\item{frac}{The extant tip of tree is n/frac}
}
\value{
result
}
\description{
data simulation
}
