% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_data_visualization.R
\name{treeplot}
\alias{treeplot}
\title{visualization of data as tree plot}
\usage{
i = 1
for (i in 1:20) {
treeplot(i)
}
}
\arguments{
\item{n}{}
}
\value{
plot
}
\description{
visualization of data as tree plot
}
