% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparision.R
\docType{data}
\name{ants}
\alias{ants}
\title{Comparision of simulated data with emperical data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 1557 columns.
}
\usage{
ants
}
\description{
Comparision of simulated data with emperical data
}
\keyword{datasets}
