% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparision.R
\name{comparision}
\alias{comparision}
\title{Comparision of simulated data with emperical data}
\usage{
comparision(datasdf = datas)
}
\arguments{
\item{datasdf}{default dataset to use for plot.}
}
\value{
p
}
\description{
Comparision of simulated data with emperical data
}
